+++ locator.php
@@ -5,6 +5,10 @@
  * See LICENSE for license details.
  */
 
+use Ulrack\CliApplication\Common\UlrackCliApplicationPackage;
 use GrizzIt\Configuration\Component\Configuration\PackageLocator;
 
-PackageLocator::registerLocation(__DIR__);
+PackageLocator::registerLocation(
+    __DIR__,
+    UlrackCliApplicationPackage::PACKAGE_NAME
+);