+++ src/Common/UlrackCliApplicationPackage.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * Copyright (C) GrizzIT, Inc. All rights reserved.
+ * See LICENSE for license details.
+ */
+
+namespace Ulrack\CliApplication\Common;
+
+interface UlrackCliApplicationPackage
+{
+    /**
+     * Package name for configuration sequencing.
+     */
+    public const PACKAGE_NAME = 'Ulrack_CliApplication';
+}