+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2020-06-14
+
+### Added
+- Ability to clear individual caches.
+
+### Fixed
+- Clearing all caches when they are not all loaded.
+
 ## 1.0.1 - 2020-06-03
 
 ### Fixed
@@ -15,5 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The initial implementation of the package.
 
 # Versions
-- [1.0.1 > Unreleased](https://github.com/ulrack/kernel/compare/1.0.1...HEAD)
+- [1.1.0 > Unreleased](https://github.com/ulrack/kernel/compare/1.1.0...HEAD)
+- [1.0.1 > 1.1.0](https://github.com/ulrack/kernel/compare/1.0.1...1.1.0)
 - [1.0.0 > 1.0.1](https://github.com/ulrack/kernel/compare/1.0.0...1.0.1)